New onRuleUpdate method implements a Hot Reload mechanism using a ReentrantReadWriteLock.
The Write Lock (Update Phase): When a rule update message arrives, acquire the writeLock() because it ensures that no transactions are being processed while destroying and recreating the Drools engines.

Availability Management: Added respondUnavailable method. It returns an ALERT_UNAVAILABLE status to NATS. This prevents the system from giving a false "Clean" response when it is actually unable to check the fraud rules.

The Swap Strategy: using a "Prepare and Swap" strategy:
    Load new rules.
    Create a temporary DroolsSessionFactory.
    Pre-fill a temporary BlockingQueue with sessions.
    Swap: Replace the global reference with the new one and dispose() the old sessions.